requirements.txt
streamlit==0.88.0
boto3==1.18.11
gspread==4.0.1
google-auth==2.6.0
Pillow==8.3.2
